2. Tasks/q. Pycaret + Pipeline/requirements.txt
streamlit==1.44.1
pandas==1.5.3
pycaret==3.3.2
numpy==1.23.5
scikit-learn==1.4.2
scipy==1.10.1
matplotlib==3.7.1
pillow==11.1.0
xgboost==1.7.6
lightgbm==3.3.5
plotly==5.24.1
joblib==1.2.0
mlflow-skinny==2.21.3
category-encoders==2.6.4
scikit-plot==0.3.7
yellowbrick==1.5